backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
httpx==0.27.2
sqlmodel==0.0.21
psycopg[binary]==3.2.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
cryptography==43.0.3
pytest==8.3.3
pytest-asyncio==0.23.8
ruff==0.6.9
mypy==1.11.2
